refactor: Footer 컴포넌트 분리 - BottomSheet 올라오는 속도 0.5초로 변경
