+++ src/app/place/components/index.ts
@@ -5,6 +5,6 @@ export { default as Tag } from './Tag';
 export { default as Tooltip } from './Tooltip';
 export { default as Info } from './Info';
 export { default as ReviewSheet } from './ReviewSheet';
-export { default as TimeLogSheet } from './TimeLogSheet';
+export { default as TimeLogSheet } from './Footer/TimeLogSheet';
 export { default as ShareSheet } from './ShareSheet';
 export { default as PlaceTopNavigationBar } from './PlaceTopNavigationBar';
